Auth Updated

+++ Tick/database/migrations/2021_04_17_062159_create_user_details_table.php
@@ -19,7 +19,7 @@ class CreateUserDetailsTable extends Migration
             $table->string('fname');
             $table->string('lname');
             $table->string('image')->nullable();
-            $table->integer('contact_num');
+            $table->bigInteger('contact_num');
             $table->date('birthdate');
             $table->enum('gender',['Male','Female']);
             $table->enum('educational_attainment',['Highschool','College']);
