+++ Tick/database/migrations/2021_04_17_062330_create_addresses_table.php
@@ -19,7 +19,7 @@ class CreateAddressesTable extends Migration
             $table->string('barangay');
             $table->string('town');
             $table->string('province');
-            $table->integer('postal_code');
+            $table->bigInteger('postal_code');
             $table->timestamps();
         });
     }
